database function get_frequent_prediction_pairs

+++ src/Domain/JackpotPlot.Domain/Models/LuckyPairResult.cs
@@ -1,3 +1,3 @@
 ﻿namespace JackpotPlot.Domain.Models;
 
-public record LuckyPairResult(int Number1, int Number2, int Frequency);
+public record LuckyPairResult(int Number1, int Number2, long Frequency);
